Update tests for nightshade feature.
